requirements.txt
psycopg2-binary==2.9.9
kafka-python==2.0.2
fastavro==1.9.4
pandas==2.1.4
PyYAML==6.0.1